+++ api/src/test/java/org/pmiops/workbench/api/CohortsControllerTest.java
@@ -281,7 +281,7 @@ public class CohortsControllerTest extends SpringTest {
   @BeforeEach
   public void setUp() {
     testMockFactory = new TestMockFactory();
-    testMockFactory.stubBufferBillingProject(billingProjectBufferService);
+    testMockFactory.stubCreateBillingProject(billingProjectBufferService);
     testMockFactory.stubCreateFcWorkspace(fireCloudService);
     DbUser user = new DbUser();
     user.setUsername(CREATOR_EMAIL);
